+++ src/apps/backend/modules/notification/internals/store/account_notification_preferences_repository.py
@@ -1,7 +1,7 @@
 from pymongo.collection import Collection
 from pymongo.errors import OperationFailure
 
-from modules.account.internal.store.account_notification_preferences_model import AccountNotificationPreferencesModel
+from modules.notification.internals.store.account_notification_preferences_model import AccountNotificationPreferencesModel
 from modules.application.repository import ApplicationRepository
 from modules.logger.logger import Logger
